Moved calculations/plotting to dedicated modules.
Moved generate_function dict to utils.py.
