+++ sage_analysis/example.py
@@ -29,8 +29,13 @@ details.
 Author: Jacob Seiler.
 """
 
-import sage_analysis.model
-import sage_analysis.plots
+from sage_analysis.model import Model
+from utils import generate_func_dict
+
+# These are contain example functions to calculate (and plot) properties such as the
+# stellar mass function, quiescent fraction, bulge fraction etc.
+import sage_analysis.example_calcs
+import sage_analysis.example_plots
 
 # Import the Data Classes that handle the different SAGE output formats.
 from sage_analysis.sage_binary import SageBinaryData
@@ -52,129 +57,6 @@ import warnings
 warnings.simplefilter(action='ignore', category=FutureWarning)
 
 
-def generate_func_dict(plot_toggles, module_name, function_prefix, keyword_args={}):
-    """
-    Generates a dictionary where the keys are the function name and the value is a list
-    containing the function itself (0th element) and keyword arguments as a dictionary
-    (1st element). All functions in the returned dictionary are expected to have the same
-    call signature for non-keyword arguments. Functions are only added when the
-    ``plot_toggles`` value is non-zero.
-
-    Functions are required to be named ``<module_prefix><function_prefix><plot_toggle_key>``
-    For example, the default calculation function are kept in the ``model.py`` module and
-    are named ``calc_<toggle>``.  E.g., ``sage_analysis.model.calc_SMF()``,
-    ``sage_analysis.model.calc_BTF()``, ``sage_analysis.model.calc_sSFR()`` etc.
-
-    Parameters
-    ----------
-
-    plot_toggles: dict, [string, int]
-        Dictionary specifying the name of each property/plot and whether the values
-        will be generated + plotted. A value of 1 denotes plotting, whilst a value of
-        0 denotes not plotting.  Entries with a value of 1 will be added to the function
-        dictionary.
-
-    module_prefix: string
-        Name of the module where the functions are located. If the functions are located
-        in this module, pass an empty string "".
-
-    function_prefix: string
-        Prefix that is added to the start of each function.
-
-    keyword_args: dict [string, dict[string, variable]], optional
-        Allows the adding of keyword aguments to the functions associated with the
-        specified plot toggle. The name of each keyword argument and associated value is
-        specified in the inner dictionary.
-
-    Returns
-    -------
-
-    func_dict: dict [string, list(function, dict[string, variable])]
-        The key of this dictionary is the name of the function.  The value is a list with
-        the 0th element being the function and the 1st element being a dictionary of
-        additional keyword arguments to be passed to the function. The inner dictionary is
-        keyed by the keyword argument names with the value specifying the keyword argument
-        value.
-
-    Examples
-    --------
-    >>> plot_toggles = {"SMF": 1}
-    >>> module_name = "sage_analysis.model"
-    >>> function_prefix = "calc_"
-    >>> generate_func_dict(plot_toggles, module_name, function_prefix) #doctest: +ELLIPSIS
-    {'calc_SMF': [<function calc_SMF at 0x...>, {}]}
-    >>> module_name = "sage_analysis.plots"
-    >>> function_prefix = "plot_"
-    >>> generate_func_dict(plot_toggles, module_name, function_prefix) #doctest: +ELLIPSIS
-    {'plot_SMF': [<function plot_SMF at 0x...>, {}]}
-
-    >>> plot_toggles = {"SMF": 1}
-    >>> module_name = "sage_analysis.plots"
-    >>> function_prefix = "plot_"
-    >>> keyword_args = {"SMF": {"plot_sub_populations": True}}
-    >>> generate_func_dict(plot_toggles, module_name, function_prefix, keyword_args) #doctest: +ELLIPSIS
-    {'plot_SMF': [<function plot_SMF at 0x...>, {'plot_sub_populations': True}]}
-
-    >>> plot_toggles = {"SMF": 1, "quiescent": 1}
-    >>> module_name = "sage_analysis.plots"
-    >>> function_prefix = "plot_"
-    >>> keyword_args = {"SMF": {"plot_sub_populations": True},
-    ...                 "quiescent": {"plot_output_format": ".pdf", "plot_sub_populations": True}}
-    >>> generate_func_dict(plot_toggles, module_name, function_prefix, keyword_args) #doctest: +ELLIPSIS
-    {'plot_SMF': [<function plot_SMF at 0x...>, {'plot_sub_populations': True}], \
-'plot_quiescent': [<function plot_quiescent at 0x...>, {'plot_output_format': '.pdf', \
-'plot_sub_populations': True}]}
-    """
-
-    # If the functions are defined in this module, then `module_name` is empty. Need to
-    # treat this differently.
-    import sys
-    if module_name == "":
-
-        # Get the name of this module.
-        module = sys.modules[__name__]
-
-    else:
-
-        # Otherwise, check if the specified module is present.
-        try:
-            module = sys.modules[module_name]
-        except KeyError:
-            msg = "Module {0} has not been imported.\nPerhaps you need to create an empty " \
-                  "`__init__.py` file to ensure your package can be imported.".format(module_name)
-            raise KeyError(msg)
-
-    # Only populate those methods that have been marked in the `plot_toggles`
-    # dictionary.
-    func_dict = {}
-    for toggle in plot_toggles.keys():
-        if plot_toggles[toggle]:
-
-            func_name = "{0}{1}".format(function_prefix, toggle)
-
-            # Be careful.  Maybe the func for a specified `plot_toggle` value wasn't
-            # added to the module.
-            try:
-                func = getattr(module, func_name)
-            except AttributeError:
-                msg = "Tried to get the func named '{0}' corresponding to " \
-                      "'plot_toggle' value '{1}'.  However, no func named '{0}' " \
-                      "could be found in '{2}' module.".format(func_name,
-                      toggle, module_prefix)
-                raise AttributeError(msg)
-
-            # We may have specified some keyword arguments for this plot toggle. Check.
-            try:
-                key_args = keyword_args[toggle]
-            except KeyError:
-                # No extra arguments for this.
-                key_args = {}
-
-            func_dict[func_name] = [func, key_args]
-
-    return func_dict
-
-
 if __name__ == "__main__":
 
     import os
@@ -186,35 +68,27 @@ if __name__ == "__main__":
     # specified if using binary output. HDF5 will automatically detect these.
     # `hdf5_snapshot` is only nedded if using HDF5 output.
 
-    model0_snapshot = 131
+    model0_snapshot = 63
     model0_IMF = "Chabrier"  # Chabrier or Salpeter.
-    model0_label = "Genesis - HDF5"  # Goes on the axis.
-    model0_sage_file = "../input/L500_N2160_genesis_hdf5.par"
-    model0_simulation = "Genesis-L500-N2160"
+    model0_label = "Mini-Millennium"  # Goes on the axis.
+    model0_sage_file = "../input/millennium.par"
+    model0_simulation = "Mini-Millennium"
     model0_first_file = 0
-    model0_last_file = 31
-
-    model1_snapshot = 131
-    model1_IMF = "Chabrier"  # Chabrier or Salpeter.
-    model1_label = "Genesis - Binary"  # Goes on the axis.
-    model1_sage_file = "../input/L500_N2160_genesis_binary.par"
-    model1_simulation = "Genesis-L500-N2160"
-    model1_first_file = 0
-    model1_last_file = 31
+    model0_last_file = 0
 
     # Then extend each of these lists for all the models that you want to plot.
     # E.g., 'IMFs = [model0_IMF, model1_IMF, ..., modelN_IMF]
-    IMFs = [model0_IMF, model1_IMF]
-    labels = [model0_label, model1_label]
-    snapshots = [model0_snapshot, model1_snapshot]
-    sage_files = [model0_sage_file, model1_sage_file]
-    simulations = [model0_simulation, model1_simulation]
-    first_files = [model0_first_file, model1_first_file]
-    last_files = [model0_last_file, model1_last_file]
+    IMFs = [model0_IMF]
+    labels = [model0_label]
+    snapshots = [model0_snapshot]
+    sage_files = [model0_sage_file]
+    simulations = [model0_simulation]
+    first_files = [model0_first_file]
+    last_files = [model0_last_file]
 
     # A couple of extra variables...
     plot_output_format    = ".png"
-    plot_output_path = "./genesis_plots"  # Will be created if path doesn't exist.
+    plot_output_path = "./plots"  # Will be created if path doesn't exist.
 
     # These toggles specify which plots you want to be made.
     plot_toggles = {"SMF"             : 1,  # Stellar mass function.
@@ -257,7 +131,7 @@ if __name__ == "__main__":
     models = []
     for model_dict in model_dicts:
 
-        my_model = sage_analysis.model.Model(model_dict)
+        my_model = Model(model_dict)
         my_model.plot_output_format = plot_output_format
 
         # Each SAGE output has a specific class written to read in the data.
@@ -285,7 +159,7 @@ if __name__ == "__main__":
         # your functions are in a different module or different function prefix, change it
         # here.
         # ALL FUNCTIONS MUST HAVE A FUNCTION SIGNATURE `func(Model, gals)`.
-        calculation_functions = generate_func_dict(plot_toggles, module_name="sage_analysis.model",
+        calculation_functions = generate_func_dict(plot_toggles, module_name="sage_analysis.example_calcs",
                                                    function_prefix="calc_")
 
         # Finally, before we calculate the properties, we need to decide how each property
@@ -333,7 +207,7 @@ if __name__ == "__main__":
 
     # Similar to the calculation functions, all of the plotting functions are in the
     # `plots.py` module and are labelled `plot_<toggle>`.
-    plot_functions = generate_func_dict(plot_toggles, module_name="sage_analysis.plots",
+    plot_functions = generate_func_dict(plot_toggles, module_name="sage_analysis.example_plots",
                                         function_prefix="plot_")
 
     # Now do the plotting.
